chore: rename to trainee
