+++ CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.8)
-project(raspicat_speak2)
+project(trainee_speak)
 
 if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
   add_compile_options(-O3 -Wall -Wextra -Wpedantic)
@@ -11,9 +11,9 @@ ament_auto_find_build_dependencies()
 
 include_directories(include)
 
-ament_auto_add_executable(raspicat_speak2_node
-  src/raspicat_speak2_core.cpp
-  src/raspicat_speak2_node.cpp
+ament_auto_add_executable(trainee_speak_node
+  src/trainee_speak_core.cpp
+  src/trainee_speak_node.cpp
 )
 
 if(BUILD_TESTING)
