+++ README.md
@@ -1,4 +1,4 @@
-# raspicat_speak2 [![build-test](https://github.com/CIT-Autonomous-Robot-Lab/raspicat_speak2/actions/workflows/build-test.yaml/badge.svg)](https://github.com/CIT-Autonomous-Robot-Lab/raspicat_speak2/actions/workflows/build-test.yaml)
+# trainee_speak [![build-test](https://github.com/CIT-Autonomous-Robot-Lab/trainee_speak/actions/workflows/build-test.yaml/badge.svg)](https://github.com/CIT-Autonomous-Robot-Lab/trainee_speak/actions/workflows/build-test.yaml)
 
 ## Overview
 [raspicat_speak](https://github.com/CIT-Autonomous-Robot-Lab/raspicat_speak)のROS 2実装です。
@@ -31,14 +31,14 @@
 ## 任意の文章を喋らせる
 
 ```
-ros2 launch raspicat_speak2 raspicat_speak2.launch.py
+ros2 launch trainee_speak trainee_speak.launch.py
 ros2 topic pub --once  /speak std_msgs/String "data: 'i am らずぱいきゃとにあん'"
 ```
 
 ## [yamlに設定した文章を喋らせる](config/speak_list.param.yaml)
 
 ```
-ros2 launch raspicat_speak2 raspicat_speak2.launch.py
+ros2 launch trainee_speak trainee_speak.launch.py
 ros2 topic pub /hoge1 std_msgs/Empty --once
 ros2 topic pub /hoge2 std_msgs/Empty --once
 ros2 topic pub /hoge3 std_msgs/Empty --once
