+++ include/trainee_speak/voice_config.hpp
