+++ src/trainee_speak_node.cpp
@@ -1,15 +1,15 @@
 // SPDX-FileCopyrightText: 2023 Tatsuhiro Ikebe <beike315@icloud.com>
 // SPDX-License-Identifier: Apache-2.0
 
-#include "raspicat_speak2/raspicat_speak2.hpp"
+#include "trainee_speak/trainee_speak.hpp"
 
 using namespace std::chrono_literals;
 
 int main(int argc, char *argv[]) {
   rclcpp::init(argc, argv);
 
-  auto node = std::make_shared<raspicat_speak2::RaspicatSpeak2>(
-      "raspicat_speak2_node", argv[1]);
+  auto node = std::make_shared<trainee_speak::RaspicatSpeak2>(
+      "trainee_speak_node", argv[1]);
 
   rclcpp::spin(node);
